requirements.txt
python-dotenv==0.20.0
aiogram==2.21
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.26.0
# библиотеки для парсинга
requests~=2.28.1
bs4==0.0.1
lxml~=4.9.1
beautifulsoup4~=4.11.1
selenium~=4.4.3
chromedriver==2.24.1
# библиотеки для времени выполнения функций
APScheduler~=3.9.1
tzlocal~=4.2
pyQiwiP2P~=2.0.6
typing~=3.7.4.3
loguru~=0.6.0